Install examples with garage
This commit makes setuptools copy the example files in the
garage package during installation.

Since the files classified as package data are required to be
in the package being installed, this commit moves examples
directory in src/garage.

It also adds a command `examples` in the garage script that
lists examples and displays their source code.